move fee_history to independent crate
