+++ crates/client/src/rpc/types.rs
@@ -4,13 +4,10 @@
 
 pub mod cfx;
 pub mod eth;
-mod fee_history;
-mod index;
 pub mod pos;
 mod provenance;
-pub mod pubsub;
 
-pub use cfx_rpc_primitives::{Bytes, U64};
+pub use cfx_rpc_primitives::{Bytes, Index, U64};
 
 pub use self::{
     cfx::{
@@ -23,6 +20,7 @@ pub use self::{
         filter::{CfxFilterChanges, CfxFilterLog, CfxRpcLogFilter, RevertTo},
         log::Log,
         pos_economics::PoSEconomics,
+        pubsub,
         receipt::Receipt,
         reward_info::RewardInfo,
         stat_on_gas_load::StatOnGasLoad,
@@ -44,8 +42,6 @@ pub use self::{
         vote_params_info::VoteParamsInfo,
         Account, CfxFeeHistory, SponsorInfo,
     },
-    fee_history::FeeHistory,
-    index::Index,
     provenance::Origin,
 };
 pub use cfx_rpc_cfx_types::{
@@ -54,3 +50,5 @@ pub use cfx_rpc_cfx_types::{
     },
     trace_filter::TraceFilter,
 };
+
+pub use cfx_rpc_eth_types::FeeHistory;
