+++ crates/rpc/rpc-cfx-types/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod address;
 pub mod epoch_number;
+mod fee_history;
 pub mod trace;
 pub mod trace_filter;
 
 pub use address::RpcAddress;
 pub use epoch_number::EpochNumber;
+pub use fee_history::CfxFeeHistory;
