+++ crates/rpc/rpc-cfx-types/src/trace.rs
@@ -367,34 +367,3 @@ impl LocalizedBlockTrace {
         })
     }
 }
-
-// #[derive(Debug, Serialize)]
-// #[serde(rename_all = "camelCase")]
-// pub struct EpochTrace {
-//     cfx_traces: Vec<LocalizedTrace>,
-//     eth_traces: Vec<EthLocalizedTrace>,
-//     mirror_address_map: HashMap<H160, RpcAddress>,
-// }
-
-// impl EpochTrace {
-//     pub fn new(
-//         cfx_traces: Vec<LocalizedTrace>, eth_traces: Vec<EthLocalizedTrace>,
-//     ) -> Self {
-//         let mut mirror_address_map = HashMap::new();
-//         for t in &cfx_traces {
-//             if let Action::Call(action) = &t.action {
-//                 if action.to.hex_address == CROSS_SPACE_CONTRACT_ADDRESS {
-//                     mirror_address_map.insert(
-//                         evm_map(action.from.hex_address).address,
-//                         action.from.clone(),
-//                     );
-//                 }
-//             }
-//         }
-//         Self {
-//             cfx_traces,
-//             eth_traces,
-//             mirror_address_map,
-//         }
-//     }
-// }
