+++ crates/rpc/rpc-eth-types/src/block.rs
@@ -306,85 +306,12 @@ impl Header {
     //     }
 }
 
-// /// Block representation with additional info.
-// pub type RichBlock = Rich<Block>;
-//
-// // /// Header representation with additional info.
-// // pub type RichHeader = Rich<Header>;
-//
-// /// Value representation with additional info
-// #[derive(Debug, Clone, PartialEq, Eq)]
-// pub struct Rich<T> {
-//     /// Standard value.
-//     pub inner: T,
-//     /// Engine-specific fields with additional description.
-//     /// Should be included directly to serialized block object.
-//     // TODO [ToDr] #[serde(skip_serializing)]
-//     pub extra_info: BTreeMap<String, String>,
-// }
-//
-// impl<T> Deref for Rich<T> {
-//     type Target = T;
-//
-//     fn deref(&self) -> &Self::Target { &self.inner }
-// }
-//
-// impl<T: Serialize> Serialize for Rich<T> {
-//     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-//     where S: Serializer {
-//         use serde_json::{to_value, Value};
-//
-//         let serialized = (to_value(&self.inner), to_value(&self.extra_info));
-//         if let (Ok(Value::Object(mut value)), Ok(Value::Object(extras))) =
-//             serialized
-//         {
-//             // join two objects
-//             value.extend(extras);
-//             // and serialize
-//             value.serialize(serializer)
-//         } else {
-//             Err(S::Error::custom(
-//                 "Unserializable structures: expected objects",
-//             ))
-//         }
-//     }
-// }
-
 #[cfg(test)]
 mod tests {
     use super::{Block, BlockTransactions};
     use crate::Bytes;
     use cfx_types::{Bloom as H2048, H160, H256, H64, U256};
-    //     use super::{Block, BlockTransactions, Header, RichBlock, RichHeader};
-    //     use ethereum_types::{Bloom as H2048, H160, H256, H64, U256};
-    //     use serde_json;
-    //     use std::collections::BTreeMap;
-    //     use v1::types::{Bytes, Transaction};
-    //
-    //     #[test]
-    //     fn test_serialize_block_transactions() {
-    //         let t = BlockTransactions::Full(vec![Transaction::default()]);
-    //         let serialized = serde_json::to_string(&t).unwrap();
-    //         assert_eq!(
-    //             serialized,
-    //
-    // r#"[{"hash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // nonce":" 0x0","blockHash":null,"blockNumber":null,"transactionIndex":
-    // null,"from":" 0x0000000000000000000000000000000000000000","to":null,"
-    // value":"0x0"," gasPrice":"0x0","gas":"0x0","input":"0x","creates":
-    // null,"raw":"0x"," publicKey":null,"chainId":null,"v":"0x0","r":"0x0",
-    // "s":"0x0","condition": null}]"#         );
-    //
-    //         let t = BlockTransactions::Hashes(vec![H256::default().into()]);
-    //         let serialized = serde_json::to_string(&t).unwrap();
-    //         assert_eq!(
-    //             serialized,
-    //
-    // r#"["0x0000000000000000000000000000000000000000000000000000000000000000"
-    // ]"#         );
-    //     }
-    //
+
     #[test]
     fn test_serialize_block() {
         let block = Block {
@@ -418,162 +345,4 @@ mod tests {
             r#"{"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","parentHash":"0x0000000000000000000000000000000000000000000000000000000000000000","sha3Uncles":"0x0000000000000000000000000000000000000000000000000000000000000000","author":"0x0000000000000000000000000000000000000000","miner":"0x0000000000000000000000000000000000000000","stateRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","transactionsRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","receiptsRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","number":"0x0","gasUsed":"0x0","gasLimit":"0x0","extraData":"0x","logsBloom":"0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","timestamp":"0x0","difficulty":"0x0","totalDifficulty":"0x0","uncles":[],"transactions":[],"size":"0x45","nonce":"0x0000000000000000","mixHash":"0x0000000000000000000000000000000000000000000000000000000000000000"}"#
         );
     }
-    //
-    //     #[test]
-    //     fn none_size_null() {
-    //         let block = Block {
-    //             hash: Some(H256::default()),
-    //             parent_hash: H256::default(),
-    //             uncles_hash: H256::default(),
-    //             author: H160::default(),
-    //             miner: H160::default(),
-    //             state_root: H256::default(),
-    //             transactions_root: H256::default(),
-    //             receipts_root: H256::default(),
-    //             number: Some(U256::default()),
-    //             gas_used: U256::default(),
-    //             gas_limit: U256::default(),
-    //             extra_data: Bytes::default(),
-    //             logs_bloom: Some(H2048::default()),
-    //             timestamp: U256::default(),
-    //             difficulty: U256::default(),
-    //             total_difficulty: Some(U256::default()),
-    //             seal_fields: vec![Bytes::default(), Bytes::default()],
-    //             base_fee_per_gas: None,
-    //             uncles: vec![],
-    //             transactions: BlockTransactions::Hashes(vec![].into()),
-    //             size: None,
-    //         };
-    //         let serialized_block = serde_json::to_string(&block).unwrap();
-    //         let rich_block = RichBlock {
-    //             inner: block,
-    //             extra_info: map![
-    //                 "mixHash".into() => format!("{:?}", H256::default()),
-    //                 "nonce".into() => format!("{:?}", H64::default())
-    //             ],
-    //         };
-    //         let serialized_rich_block =
-    // serde_json::to_string(&rich_block).unwrap();
-    //
-    //         assert_eq!(
-    //             serialized_block,
-    //
-    // r#"{"hash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // parentHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // sha3Uncles":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // author": "0x0000000000000000000000000000000000000000","miner":"
-    // 0x0000000000000000000000000000000000000000","stateRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // transactionsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // receiptsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // number": "0x0","gasUsed":"0x0","gasLimit":"0x0","extraData":"0x","
-    // logsBloom":"
-    // 0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-    // ,"timestamp":"0x0","difficulty":"0x0","totalDifficulty":"0x0","
-    // sealFields":[" 0x","0x"],"uncles":[],"transactions":[],"size":null}"#
-    // );         assert_eq!(
-    //             serialized_rich_block,
-    //
-    // r#"{"author":"0x0000000000000000000000000000000000000000","difficulty":"
-    // 0x0", "extraData":"0x","gasLimit":"0x0","gasUsed":"0x0","hash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // logsBloom":"
-    // 0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-    // ,"miner":"0x0000000000000000000000000000000000000000","mixHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // nonce":" 0x0000000000000000","number":"0x0","parentHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // receiptsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // sealFields":["0x","0x"],"sha3Uncles":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // size": null,"stateRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // timestamp":"0x0","totalDifficulty":"0x0","transactions":[],"
-    // transactionsRoot"
-    // :"0x0000000000000000000000000000000000000000000000000000000000000000","
-    // uncles":[]}"#         );
-    //     }
-    //
-    //     #[test]
-    //     fn test_serialize_header() {
-    //         let header = Header {
-    //             hash: Some(H256::default()),
-    //             parent_hash: H256::default(),
-    //             uncles_hash: H256::default(),
-    //             author: H160::default(),
-    //             miner: H160::default(),
-    //             state_root: H256::default(),
-    //             transactions_root: H256::default(),
-    //             receipts_root: H256::default(),
-    //             number: Some(U256::default()),
-    //             gas_used: U256::default(),
-    //             gas_limit: U256::default(),
-    //             extra_data: Bytes::default(),
-    //             logs_bloom: H2048::default(),
-    //             timestamp: U256::default(),
-    //             difficulty: U256::default(),
-    //             seal_fields: vec![Bytes::default(), Bytes::default()],
-    //             base_fee_per_gas: None,
-    //             size: Some(69.into()),
-    //         };
-    //         let serialized_header = serde_json::to_string(&header).unwrap();
-    //         let rich_header = RichHeader {
-    //             inner: header,
-    //             extra_info: map![
-    //                 "mixHash".into() => format!("{:?}", H256::default()),
-    //                 "nonce".into() => format!("{:?}", H64::default())
-    //             ],
-    //         };
-    //         let serialized_rich_header =
-    // serde_json::to_string(&rich_header).unwrap();
-    //
-    //         assert_eq!(
-    //             serialized_header,
-    //
-    // r#"{"hash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // parentHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // sha3Uncles":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // author": "0x0000000000000000000000000000000000000000","miner":"
-    // 0x0000000000000000000000000000000000000000","stateRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // transactionsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // receiptsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // number": "0x0","gasUsed":"0x0","gasLimit":"0x0","extraData":"0x","
-    // logsBloom":"
-    // 0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-    // ,"timestamp":"0x0","difficulty":"0x0","sealFields":["0x","0x"],"size":"
-    // 0x45" }"#         );
-    //         assert_eq!(
-    //             serialized_rich_header,
-    //
-    // r#"{"author":"0x0000000000000000000000000000000000000000","difficulty":"
-    // 0x0", "extraData":"0x","gasLimit":"0x0","gasUsed":"0x0","hash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // logsBloom":"
-    // 0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-    // ,"miner":"0x0000000000000000000000000000000000000000","mixHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // nonce":" 0x0000000000000000","number":"0x0","parentHash":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // receiptsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // sealFields":["0x","0x"],"sha3Uncles":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // size":" 0x45","stateRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000","
-    // timestamp":"0x0","transactionsRoot":"
-    // 0x0000000000000000000000000000000000000000000000000000000000000000"}"#
-    //         );
-    //     }
 }
