+++ crates/rpc/rpc-eth-types/src/fee_history.rs
@@ -1,13 +1,13 @@
+use serde::Serialize;
 use std::collections::VecDeque;
 
 use cfx_parameters::block::{
     cspace_block_gas_limit_after_cip1559,
     espace_block_gas_limit_of_enabled_block,
 };
+use cfx_rpc_cfx_types::CfxFeeHistory;
 use cfx_types::{Space, SpaceMap, U256};
 use primitives::{transaction::SignedTransaction, BlockHeader};
-
-use super::CfxFeeHistory;
 
 #[derive(Serialize, Debug, Default)]
 #[serde(rename_all = "camelCase")]
