+++ crates/rpc/rpc-eth-types/src/lib.rs
@@ -2,6 +2,7 @@ mod block;
 mod block_number;
 mod errors;
 pub mod eth_pubsub;
+mod fee_history;
 mod filter;
 mod log;
 mod receipt;
@@ -17,6 +18,7 @@ pub use block_number::BlockNumber;
 pub use cfx_rpc_primitives::{Bytes, U64};
 pub use errors::Error;
 pub use eth_pubsub::*;
+pub use fee_history::FeeHistory;
 pub use filter::*;
 pub use log::Log;
 pub use receipt::Receipt;
