+++ crates/rpc/rpc-primitives/src/lib.rs
@@ -1,8 +1,10 @@
 mod bytes;
+mod index;
 mod variadic_u64;
 mod variadic_value;
 
 pub use bytes::Bytes;
+pub use index::Index;
 pub use variadic_u64::U64;
 pub use variadic_value::VariadicValue;
